refactor: replace AST-based eval.rs with HIR-based evaluation system
- Replace eval.rs with hir_eval.rs as primary evaluation module
- Move Value, Environment, and BuiltinFn types to new eval module
- Update trbi.rs to use HIR-only evaluation without AST fallback
- Add compatibility eval_expr function for legacy AST-based code
- Remove duplicate hir_eval module and old eval_old backup
- Simplify interpreter to use only HIR evaluation path
- Update lib.rs exports to use consolidated eval module

Note: Some integration tests fail due to missing builtin functions in HIR system
Core HIR evaluation functionality is working correctly
