+++ src/eval.rs
@@ -1,5 +1,6 @@
-use tribute_ast::ast::{Expr, Identifier, Spanned};
 use crate::builtins;
+use tribute_ast::{ast::Identifier, Spanned};
+use tribute_hir::hir::{Expr, HirExpr, HirFunction, HirProgram, Literal, Pattern};
 use std::collections::HashMap;
 
 type Error = Box<dyn std::error::Error + 'static>;
@@ -12,7 +13,7 @@ pub enum Value {
     Number(i64),
     String(String),
     List(Vec<Value>),
-    Fn(Identifier, Vec<Identifier>, Vec<Spanned<Expr>>),
+    Fn(Identifier, Vec<Identifier>, Vec<Spanned<tribute_ast::ast::Expr>>),
     BuiltinFn(&'static str, BuiltinFn),
 }
 
@@ -76,139 +77,606 @@ impl Environment<'_> {
     }
 }
 
-pub fn eval_expr(env: &mut Environment<'_>, expr: &Expr) -> Result<Value, Error> {
-    use Expr::*;
-    Ok(match expr {
-        Number(n) => Value::Number(*n),
-        String(s) => Value::String(s.clone()),
-        Identifier(ident) => env.lookup(ident)?.clone(),
-        List(exprs) => {
-            let Some((head, rest)) = exprs.split_first() else {
-                return Ok(Value::Unit);
-            };
-            let head = match &head.0 {
-                // TODO: this must have been checked in the compile time
-                Number(_) => return Err("expected function, got number".into()),
-                String(_) => return Err("expected function, got string".into()),
-                Identifier(ident) => match ident.as_str() {
-                    "let" => {
-                        return handle_let(env, rest);
+// Context for HIR evaluation with access to the program
+pub struct HirEvalContext<'db> {
+    pub db: &'db dyn salsa::Database,
+    pub program: HirProgram<'db>,
+}
+
+// Main evaluation function that uses HIR-based evaluation
+pub fn eval_expr(env: &mut Environment<'_>, expr: &tribute_ast::ast::Expr) -> Result<Value, Error> {
+    // For compatibility with existing AST-based code, we need to parse and lower to HIR
+    // This is a transitional function and should be phased out in favor of direct HIR evaluation
+    use tribute_ast::TributeDatabaseImpl;
+    use tribute_ast::ast::Expr;
+    use salsa::Database;
+    
+    match expr {
+        // Handle function definitions by storing them in the environment
+        Expr::List(exprs) if !exprs.is_empty() => {
+            if let (Expr::Identifier(fn_name), _) = &exprs[0] {
+                if fn_name == "fn" && exprs.len() >= 3 {
+                    // This is a function definition - store it in the environment
+                    if let (Expr::List(sig_parts), _) = &exprs[1] {
+                        if !sig_parts.is_empty() {
+                            if let (Expr::Identifier(name), _) = &sig_parts[0] {
+                                let param_names: Result<Vec<String>, Error> = sig_parts
+                                    .iter()
+                                    .skip(1)
+                                    .map(|(expr, _)| match expr {
+                                        Expr::Identifier(id) => Ok(id.clone()),
+                                        _ => Err("function parameter must be an identifier".into()),
+                                    })
+                                    .collect();
+                                
+                                let param_names = param_names?;
+                                let body_exprs = exprs[2..].to_vec();
+                                let function_value = Value::Fn(name.clone(), param_names, body_exprs);
+                                env.bind(name.clone(), function_value);
+                                return Ok(Value::Unit);
+                            }
+                        }
                     }
-                    "fn" => {
-                        return handle_fn(env, rest);
-                    }
-                    "match" => {
-                        return handle_match(env, rest);
-                    }
-                    _ => env.lookup(ident)?.clone(),
-                },
-                _ => eval_expr(env, &head.0)?,
-            };
-            match head {
-                Value::Fn(_, params, body) => {
-                    let bindings = params
-                        .iter()
-                        .zip(rest.iter())
-                        .map(|(param, arg)| Ok((param.clone(), eval_expr(env, &arg.0)?)))
-                        .collect::<Result<Vec<_>, Error>>()?;
-                    let mut child_env = env.child(bindings);
-                    let mut result = Value::Unit;
-                    for expr in body {
-                        result = eval_expr(&mut child_env, &expr.0)?;
-                    }
-                    result
                 }
-                Value::BuiltinFn(_, f) => {
-                    let args = rest
-                        .iter()
-                        .map(|expr| eval_expr(env, &expr.0))
-                        .collect::<Result<Vec<_>, _>>()?;
-                    f(&args)?
+                
+                // Handle let bindings
+                if fn_name == "let" && exprs.len() == 3 {
+                    if let (Expr::Identifier(ident), _) = &exprs[1] {
+                        let value = eval_expr(env, &exprs[2].0)?;
+                        env.bind(ident.clone(), value.clone());
+                        return Ok(value);
+                    }
                 }
-                _ => unimplemented!(),
             }
+        }
+        _ => {}
+    }
+    
+    // For other expressions, try HIR evaluation
+    TributeDatabaseImpl::default().attach(|db| {
+        // Create a temporary source for the single expression
+        let source = format!("(fn (main) {})", format_expr_as_source(expr));
+        
+        match crate::eval_with_hir(db, "temp.trb", &source) {
+            Ok(result) => Ok(result),
+            Err(e) => Err(format!("HIR evaluation failed: {}", e).into())
         }
     })
 }
 
-fn handle_let(env: &mut Environment<'_>, rest: &[Spanned<Expr>]) -> Result<Value, Error> {
-    let [(Expr::Identifier(ident), _), value] = rest else {
-        return Err("expected identifier and value".into());
-    };
-    let value = eval_expr(env, &value.0)?;
-    env.bind(ident.clone(), value);
-    Ok(Value::Unit)
+// Helper function to format an AST expression back to source code
+fn format_expr_as_source(expr: &tribute_ast::ast::Expr) -> String {
+    use tribute_ast::ast::Expr;
+    match expr {
+        Expr::Number(n) => n.to_string(),
+        Expr::String(s) => format!("\"{}\"", s),
+        Expr::Identifier(id) => id.clone(),
+        Expr::List(exprs) => {
+            let formatted_exprs: Vec<String> = exprs.iter()
+                .map(|(expr, _)| format_expr_as_source(expr))
+                .collect();
+            format!("({})", formatted_exprs.join(" "))
+        }
+    }
 }
 
-fn handle_fn(env: &mut Environment<'_>, rest: &[Spanned<Expr>]) -> Result<Value, Error> {
-    if rest.len() < 2 {
-        return Err("fn requires signature and body".into());
+pub fn eval_hir_program<'db>(
+    db: &'db dyn salsa::Database,
+    env: &mut Environment<'_>,
+    program: HirProgram<'db>,
+) -> Result<Value, Error> {
+    let functions = program.functions(db);
+
+    // Bind all functions to environment first
+    for (name, func) in &functions {
+        let func_value = Value::Fn(
+            name.clone(),
+            func.params(db).to_vec(),
+            vec![], // HIR functions store their body differently
+        );
+        env.bind(name.clone(), func_value);
     }
 
-    let (sig, body_exprs) = rest.split_first().unwrap();
-    let Expr::List(sig_parts) = &sig.0 else {
-        return Err("expected signature list".into());
-    };
+    let context = HirEvalContext { db, program };
 
-    let Some(((Expr::Identifier(name), _), params)) = sig_parts.split_first() else {
-        return Err("expected name and parameters".into());
-    };
+    // If there's a main function, call it
+    if let Some(main_name) = program.main(db) {
+        if let Some(main_func) = functions.get(&main_name) {
+            eval_hir_function_body_with_context(&context, *main_func, vec![])
+        } else {
+            Err(format!("main function '{}' not found", main_name).into())
+        }
+    } else {
+        // No main function, just return unit
+        Ok(Value::Unit)
+    }
+}
 
-    let param_names: Result<Vec<Identifier>, Error> = params
+pub fn eval_hir_function_body<'db>(
+    db: &'db dyn salsa::Database,
+    program: HirProgram<'db>,
+    env: &mut Environment<'_>,
+    func: HirFunction<'db>,
+    args: Vec<Value>,
+) -> Result<Value, Error> {
+    let params = func.params(db);
+    if args.len() != params.len() {
+        return Err(format!(
+            "function {} expects {} arguments, got {}",
+            func.name(db),
+            params.len(),
+            args.len()
+        )
+        .into());
+    }
+
+    // Create bindings for function parameters
+    let bindings: Vec<(std::string::String, Value)> = params
         .iter()
-        .map(|(expr, _)| match expr {
-            Expr::Identifier(id) => Ok(id.clone()),
-            _ => Err("function parameter must be an identifier".into()),
-        })
+        .zip(args)
+        .map(|(param, value)| (param.clone(), value))
         .collect();
 
-    let param_names = param_names?;
-    let function_value = Value::Fn(name.clone(), param_names, body_exprs.to_vec());
-    env.bind(name.clone(), function_value);
+    let mut child_env = env.child(bindings);
 
-    Ok(Value::Unit)
-}
+    // Create context for evaluation
+    let context = HirEvalContext { db, program };
 
-fn handle_match(env: &mut Environment<'_>, rest: &[Spanned<Expr>]) -> Result<Value, Error> {
-    if rest.is_empty() {
-        return Err("match requires at least one argument".into());
+    // Evaluate all expressions in the function body
+    let body = func.body(db);
+    let mut result = Value::Unit;
+    for expr in body {
+        result = eval_hir_expr(&context, &mut child_env, expr)?;
     }
 
-    let (value_expr, cases) = rest.split_first().unwrap();
-    let value = eval_expr(env, &value_expr.0)?;
+    Ok(result)
+}
 
-    for case in cases {
-        let Expr::List(case_parts) = &case.0 else {
-            continue; // Skip non-list cases
-        };
+pub fn eval_hir_function_body_with_context<'db>(
+    context: &HirEvalContext<'db>,
+    func: HirFunction<'db>,
+    args: Vec<Value>,
+) -> Result<Value, Error> {
+    let params = func.params(context.db);
+    if args.len() != params.len() {
+        return Err(format!(
+            "function {} expects {} arguments, got {}",
+            func.name(context.db),
+            params.len(),
+            args.len()
+        )
+        .into());
+    }
 
-        if case_parts.len() != 3 {
-            continue; // Skip malformed cases
-        }
+    // Create bindings for function parameters
+    let bindings: Vec<(Identifier, Value)> = params
+        .iter()
+        .zip(args)
+        .map(|(param, value)| (param.clone(), value))
+        .collect();
 
-        let case_keyword = &case_parts[0].0;
-        let pattern = &case_parts[1].0;
-        let result_expr = &case_parts[2].0;
+    // Create new environment with all functions bound
+    let mut env = Environment::toplevel();
 
-        if let Expr::Identifier(keyword) = case_keyword {
-            if keyword == "case" {
-                let pattern_value = eval_expr(env, pattern)?;
-                if values_match(&value, &pattern_value) {
-                    return eval_expr(env, result_expr);
+    // Bind all functions in the program
+    let functions = context.program.functions(context.db);
+    for (name, hir_func) in &functions {
+        let func_value = Value::Fn(
+            name.clone(),
+            hir_func.params(context.db),
+            vec![], // HIR doesn't store the original AST body
+        );
+        env.bind(name.clone(), func_value);
+    }
+
+    let mut child_env = env.child(bindings);
+
+    // Evaluate all expressions in the function body
+    let body = func.body(context.db);
+    let mut result = Value::Unit;
+    for expr in body {
+        result = eval_hir_expr(context, &mut child_env, expr)?;
+    }
+
+    Ok(result)
+}
+
+pub fn eval_hir_expr<'db>(
+    context: &HirEvalContext<'db>,
+    env: &mut Environment<'_>,
+    expr: HirExpr<'db>,
+) -> Result<Value, Error> {
+    eval_spanned_expr(
+        context,
+        env,
+        (expr.expr(context.db).clone(), expr.span(context.db)),
+    )
+}
+
+fn eval_spanned_expr<'db>(
+    context: &HirEvalContext<'db>,
+    env: &mut Environment<'_>,
+    (expr, _span): Spanned<Expr>,
+) -> Result<Value, Error> {
+    use Expr::*;
+
+    match expr {
+        Number(n) => Ok(Value::Number(n)),
+        String(s) => Ok(Value::String(s)),
+        Variable(name) => env.lookup(&name).cloned(),
+        Call { func, args } => {
+            let func_expr = eval_spanned_expr(context, env, *func)?;
+            match func_expr {
+                Value::Fn(name, params, _) => {
+                    let arg_values: Result<Vec<Value>, Error> = args
+                        .into_iter()
+                        .map(|arg| eval_spanned_expr(context, env, arg))
+                        .collect();
+                    let arg_values = arg_values?;
+
+                    // Create child environment with parameter bindings
+                    if params.len() != arg_values.len() {
+                        return Err(format!(
+                            "function {} expects {} arguments, got {}",
+                            name,
+                            params.len(),
+                            arg_values.len()
+                        )
+                        .into());
+                    }
+
+                    let bindings: Vec<(Identifier, Value)> =
+                        params.into_iter().zip(arg_values.iter().cloned()).collect();
+
+                    let _child_env = env.child(bindings);
+
+                    // Find the function in the program and evaluate its body
+                    let functions = context.program.functions(context.db);
+                    if let Some((_, func)) = functions.iter().find(|(fname, _)| **fname == name) {
+                        eval_hir_function_body_with_context(context, *func, arg_values)
+                    } else {
+                        Err(format!("Function '{}' not found", name).into())
+                    }
                 }
+                Value::BuiltinFn(_, f) => {
+                    let arg_values: Result<Vec<Value>, Error> = args
+                        .into_iter()
+                        .map(|arg| eval_spanned_expr(context, env, arg))
+                        .collect();
+                    f(&arg_values?)
+                }
+                _ => Err(format!("not a function: {:?}", func_expr).into()),
             }
         }
+        Let { var, value } => {
+            let val = eval_spanned_expr(context, env, *value)?;
+            env.bind(var, val.clone());
+            Ok(val)
+        }
+        Match { expr, cases } => {
+            let value = eval_spanned_expr(context, env, *expr)?;
+            for case in cases {
+                if let Some(bindings) = match_pattern(&value, &case.pattern) {
+                    let mut child_env = env.child(bindings);
+                    return eval_spanned_expr(context, &mut child_env, case.body);
+                }
+            }
+            Err("no matching case found".into())
+        }
+        Builtin { name, args } => eval_builtin_by_name(context, env, &name, args),
+        Block(exprs) => {
+            let mut result = Value::Unit;
+            for expr in exprs {
+                result = eval_spanned_expr(context, env, expr)?;
+            }
+            Ok(result)
+        }
     }
-
-    Err("no matching case found".into())
 }
 
-fn values_match(a: &Value, b: &Value) -> bool {
-    match (a, b) {
-        (Value::Number(a), Value::Number(b)) => a == b,
-        (Value::String(a), Value::String(b)) => a == b,
-        (Value::Unit, Value::Unit) => true,
-        _ => false,
+fn eval_builtin_by_name<'db>(
+    context: &HirEvalContext<'db>,
+    env: &mut Environment<'_>,
+    name: &str,
+    args: Vec<Spanned<Expr>>,
+) -> Result<Value, Error> {
+    let arg_values: Result<Vec<Value>, Error> = args
+        .into_iter()
+        .map(|arg| eval_spanned_expr(context, env, arg))
+        .collect();
+    let arg_values = arg_values?;
+
+    match name {
+        "+" => {
+            if arg_values.len() != 2 {
+                return Err("+ requires exactly 2 arguments".into());
+            }
+            match (&arg_values[0], &arg_values[1]) {
+                (Value::Number(a), Value::Number(b)) => Ok(Value::Number(a + b)),
+                _ => Err("+ requires numbers".into()),
+            }
+        }
+        "-" => {
+            if arg_values.len() != 2 {
+                return Err("- requires exactly 2 arguments".into());
+            }
+            match (&arg_values[0], &arg_values[1]) {
+                (Value::Number(a), Value::Number(b)) => Ok(Value::Number(a - b)),
+                _ => Err("- requires numbers".into()),
+            }
+        }
+        "*" => {
+            if arg_values.len() != 2 {
+                return Err("* requires exactly 2 arguments".into());
+            }
+            match (&arg_values[0], &arg_values[1]) {
+                (Value::Number(a), Value::Number(b)) => Ok(Value::Number(a * b)),
+                _ => Err("* requires numbers".into()),
+            }
+        }
+        "/" => {
+            if arg_values.len() != 2 {
+                return Err("/ requires exactly 2 arguments".into());
+            }
+            match (&arg_values[0], &arg_values[1]) {
+                (Value::Number(a), Value::Number(b)) => {
+                    if *b == 0 {
+                        Err("division by zero".into())
+                    } else {
+                        Ok(Value::Number(a / b))
+                    }
+                }
+                _ => Err("/ requires numbers".into()),
+            }
+        }
+        "print_line" => {
+            if let Some(builtin_fn) = builtins::BUILTINS.get("print_line") {
+                if let Value::BuiltinFn(_, f) = builtin_fn {
+                    f(&arg_values)
+                } else {
+                    Err("print_line is not a builtin function".into())
+                }
+            } else {
+                Err("print_line not found".into())
+            }
+        }
+        "input_line" => {
+            if let Some(builtin_fn) = builtins::BUILTINS.get("input_line") {
+                if let Value::BuiltinFn(_, f) = builtin_fn {
+                    f(&arg_values)
+                } else {
+                    Err("input_line is not a builtin function".into())
+                }
+            } else {
+                Err("input_line not found".into())
+            }
+        }
+        _ => Err(format!("unknown builtin function: {}", name).into()),
     }
 }
 
+fn match_pattern(value: &Value, pattern: &Pattern) -> Option<Vec<(std::string::String, Value)>> {
+    match pattern {
+        Pattern::Literal(lit) => {
+            // Compare literal values
+            match (lit, value) {
+                (Literal::Number(a), Value::Number(b)) if a == b => Some(vec![]),
+                (Literal::String(a), Value::String(b)) if a == b => Some(vec![]),
+                _ => None,
+            }
+        }
+        Pattern::Variable(name) => {
+            // Variable patterns always match and bind
+            Some(vec![(name.clone(), value.clone())])
+        }
+        Pattern::Wildcard => {
+            // Wildcard always matches without binding
+            Some(vec![])
+        }
+        Pattern::List(patterns) => {
+            // List pattern matching
+            match value {
+                Value::List(values) => {
+                    // Check if we have a rest pattern
+                    if let Some(last_pattern) = patterns.last() {
+                        if matches!(last_pattern, Pattern::Rest(_)) {
+                            return match_list_with_rest(values, patterns);
+                        }
+                    }
+
+                    // Exact length matching
+                    if values.len() != patterns.len() {
+                        return None;
+                    }
+
+                    let mut bindings = Vec::new();
+                    for (value, pattern) in values.iter().zip(patterns.iter()) {
+                        if let Some(mut pattern_bindings) = match_pattern(value, pattern) {
+                            bindings.append(&mut pattern_bindings);
+                        } else {
+                            return None;
+                        }
+                    }
+                    Some(bindings)
+                }
+                _ => None,
+            }
+        }
+        Pattern::Rest(_) => {
+            // Rest patterns should only appear in list contexts
+            None
+        }
+    }
+}
+
+fn match_list_with_rest(
+    values: &[Value],
+    patterns: &[Pattern],
+) -> Option<Vec<(std::string::String, Value)>> {
+    if patterns.is_empty() {
+        return None;
+    }
+
+    let (rest_pattern, head_patterns) = patterns.split_last().unwrap();
+
+    if let Pattern::Rest(rest_name) = rest_pattern {
+        // Must have at least as many values as head patterns
+        if values.len() < head_patterns.len() {
+            return None;
+        }
+
+        let mut bindings = Vec::new();
+
+        // Match head patterns
+        for (value, pattern) in values
+            .iter()
+            .take(head_patterns.len())
+            .zip(head_patterns.iter())
+        {
+            if let Some(mut pattern_bindings) = match_pattern(value, pattern) {
+                bindings.append(&mut pattern_bindings);
+            } else {
+                return None;
+            }
+        }
+
+        // Bind rest values
+        let rest_values = values[head_patterns.len()..].to_vec();
+        bindings.push((rest_name.clone(), Value::List(rest_values)));
+
+        Some(bindings)
+    } else {
+        None
+    }
+}
+
+#[cfg(test)]
+mod tests {
+    use super::*;
+    use tribute_ast::TributeDatabaseImpl;
+    use salsa::Database;
+
+    #[test]
+    fn test_hir_arithmetic() {
+        TributeDatabaseImpl::default().attach(|db| {
+            // Test simple arithmetic expression wrapped in a main function
+            let source = r#"(fn (main) (+ 1 2))"#;
+            match crate::eval_with_hir(db, "test.trb", source) {
+                Ok(Value::Number(3)) => {}
+                Ok(other) => panic!("Expected Number(3), got {:?}", other),
+                Err(e) => panic!("HIR evaluation failed: {}", e),
+            }
+        });
+    }
+
+    #[test]
+    fn test_hir_print_line() {
+        TributeDatabaseImpl::default().attach(|db| {
+            // Test print_line builtin wrapped in a main function
+            let source = r#"(fn (main) (print_line "Hello HIR"))"#;
+            match crate::eval_with_hir(db, "test.trb", source) {
+                Ok(Value::Unit) => {}
+                Ok(other) => panic!("Expected Unit, got {:?}", other),
+                Err(e) => panic!("HIR evaluation failed: {}", e),
+            }
+        });
+    }
+
+    #[test]
+    fn test_hir_nested_arithmetic() {
+        TributeDatabaseImpl::default().attach(|db| {
+            // Test nested arithmetic wrapped in a main function
+            let source = r#"(fn (main) (+ (* 2 3) (/ 8 2)))"#;
+            match crate::eval_with_hir(db, "test.trb", source) {
+                Ok(Value::Number(10)) => {}
+                Ok(other) => panic!("Expected Number(10), got {:?}", other),
+                Err(e) => panic!("HIR evaluation failed: {}", e),
+            }
+        });
+    }
+
+    #[test]
+    fn test_hir_let_binding() {
+        TributeDatabaseImpl::default().attach(|db| {
+            // Test let binding without body
+            let source = r#"(fn (main) (let x 42) x)"#;
+            match crate::eval_with_hir(db, "test.trb", source) {
+                Ok(Value::Number(42)) => {}
+                Ok(other) => panic!("Expected Number(42), got {:?}", other),
+                Err(e) => panic!("HIR evaluation failed: {}", e),
+            }
+        });
+    }
+
+    #[test]
+    fn test_hir_pattern_matching() {
+        TributeDatabaseImpl::default().attach(|db| {
+            // Test pattern matching
+            let source = r#"
+                (fn (test_number n)
+                  (match n
+                    (case 0 "zero")
+                    (case 1 "one")
+                    (case _ "other")))
+                (fn (main) (test_number 0))
+            "#;
+            match crate::eval_with_hir(db, "test.trb", source) {
+                Ok(Value::String(s)) if s == "zero" => {}
+                Ok(other) => panic!("Expected String(\"zero\"), got {:?}", other),
+                Err(e) => panic!("HIR evaluation failed: {}", e),
+            }
+        });
+    }
+
+    #[test]
+    fn test_hir_user_defined_function() {
+        TributeDatabaseImpl::default().attach(|db| {
+            // Test user-defined function call
+            let source = r#"
+                (fn (add x y) (+ x y))
+                (fn (main) (add 10 20))
+            "#;
+            match crate::eval_with_hir(db, "test.trb", source) {
+                Ok(Value::Number(30)) => {}
+                Ok(other) => panic!("Expected Number(30), got {:?}", other),
+                Err(e) => panic!("HIR evaluation failed: {}", e),
+            }
+        });
+    }
+
+    #[test]
+    fn test_hir_functions_calling_each_other() {
+        TributeDatabaseImpl::default().attach(|db| {
+            // Test functions calling each other
+            let source = r#"
+                (fn (double x) (* x 2))
+                (fn (add_and_double x y) (double (+ x y)))
+                (fn (main) (add_and_double 5 10))
+            "#;
+            match crate::eval_with_hir(db, "test.trb", source) {
+                Ok(Value::Number(30)) => {}
+                Ok(other) => panic!("Expected Number(30), got {:?}", other),
+                Err(e) => panic!("HIR evaluation failed: {}", e),
+            }
+        });
+    }
+
+    #[test]
+    fn test_hir_recursive_function() {
+        TributeDatabaseImpl::default().attach(|db| {
+            // Test recursive function (factorial)
+            let source = r#"
+                (fn (factorial n)
+                  (match n
+                    (case 0 1)
+                    (case _ (* n (factorial (- n 1))))))
+                (fn (main) (factorial 5))
+            "#;
+            match crate::eval_with_hir(db, "test.trb", source) {
+                Ok(Value::Number(120)) => {}
+                Ok(other) => panic!("Expected Number(120), got {:?}", other),
+                Err(e) => panic!("HIR evaluation failed: {}", e),
+            }
+        });
+    }
+}
